#67 refactor: openai 관련 패키지 구조 변경
